All views working and Customer design in progress

+++ src/Core/Ahmynar_Application/DTOs/Product/IProductDto.cs
@@ -11,7 +11,7 @@ namespace Ahmynar_Application.DTOs.Product
         public string Description { get; set; }
         public Ahmynar_Domain.Enums.GroupProduct GroupProduct { get; set; }
         public float PurchasePrice { get; set; }
-        public ushort Quantity { get; set; }
+        public int Quantity { get; set; }
         public string? Unit { get; set; }
         public float SalePrice { get; set; }
         public string? Obs { get; set; }
