+++ src/Core/Ahmynar_Application/DTOs/Product/UpdateProductDto.cs
@@ -12,7 +12,7 @@ namespace Ahmynar_Application.DTOs.Product
         public string Description { get; set; } = default!;
         public Ahmynar_Domain.Enums.GroupProduct GroupProduct { get; set; }
         public float PurchasePrice { get; set; }
-        public ushort Quantity { get; set; }
+        public int Quantity { get; set; }
         public string? Unit { get; set; }
         public float SalePrice { get; set; }
         public string? Obs { get; set; }
