+++ src/Core/Ahmynar_Domain/Product.cs
@@ -8,7 +8,7 @@ namespace Ahmynar_Domain
         public string Description { get; set; }
         public GroupProduct GroupProduct { get; set; }
         public float PurchasePrice { get; set; }
-        public ushort Quantity { get; set; }
+        public int Quantity { get; set; }
         public string? Unit { get; set; }
         public float SalePrice { get; set; }
         public string? Obs { get; set; }
